% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{detach_all}
\alias{detach_all}
\title{Detach All Elements}
\usage{
detach_all(except)
}
\arguments{
\item{except}{vector; elements that should be excluded from detaching}
}
\description{
This function will detach all elements except th base R packages:
.GlobalEnv', 'package:stats', 'package:graphics', 'package:grDevices', 'package:utils',
'package:datasets', 'package:methods', 'Autoloads', 'package:base', 'tools:rstudio'
_Caution: This function will also detach rootdetectR_
}
\examples{
# detach_all()
}
