% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{inspect_root_norm}
\alias{inspect_root_norm}
\title{Visualize and Inspect Rootdetection Data}
\usage{
inspect_root_norm(root_norm, plot = TRUE, output = TRUE)
}
\arguments{
\item{root_norm}{data.frame;  normalized Rootdetection data set}

\item{plot}{logical; if TRUE summary plot is produced}

\item{output}{logical; if TRUE summary is printed in R console}
}
\value{
plot and/or console output
}
\description{
The function produces a summary of all lines and conditions of a normalized rootdetection dataset.
It visualizes a summary in a matrix styled plot and gives a description of data in the R console.
}
\examples{
# normalize root_output
root_norm <- norm_10mm_standard(root_output)

# data summary
inspect_root_norm(root_norm)
}
