% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{interaction_twofacaov}
\alias{interaction_twofacaov}
\title{Pairwise Two Way ANOVA of Treatment (Interaction) Effects}
\usage{
interaction_twofacaov(root_norm, grouping_var1 = "Factor1",
  grouping_var2 = "Factor2", dependend_var = "LengthMM",
  control = "20", label_delim = ";", draw_out = F,
  file_base = "2fac_ANOVA_BH_corrected", p_value_size = 0.8)
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{control}{string; name of the Factor2 control condition}

\item{label_delim}{character; defin how Factor1 and Factor2 are seperated in Label}

\item{draw_out}{logical; If TRUE Matrix containg p-values is plotted in pdf file}

\item{file_base}{string; file name of the pdf output (is needed if draw_out = T)}

\item{p_value_size}{numeric; font siz of the p-values printed in pdf file}
}
\value{
list; matrices containg p-values for pairwise two way ANOVA for each Factor1 per Factor2 control treatment effect
}
\description{
The function performes a two way ANOVA. It compares the treatment effect of Factor2 control to Factor2 treatment for every Factor1 (pairwise). P-values are corrected using Benjamini-Hochberg procedure.
}
\examples{
# get data.frame containg p-values for pairwise two way ANOVA

root_test_norm <- norm_10mm_standard(root_output)
interaction_twofacaov(root_test_norm, control = '20', label_delim = ';', draw_out = F)

# get data.frame and plot as pdf output

root_test_norm <- norm_10mm_standard(root_output)
interaction_twofacaov(root_test_norm, control = '20', label_delim = ';', draw_out = T, file_base = '2fac_ANOVA_BH_corrected')
# change size of printed p-values
interaction_twofacaov(root_test_norm, control = '20', label_delim = ';', draw_out = T, file_base = '2fac_ANOVA_BH_corrected', p_value_size = 1)
# function creates a pdf file 2fac_ANOVA_BH_corrected_28.pdf
}
