% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{is.root_detection_output}
\alias{is.root_detection_output}
\title{Test Input data.frame for Rootdetection standard}
\usage{
is.root_detection_output(root_output)
}
\arguments{
\item{root_output}{data.frame; output *.csv from Rootdetection}
}
\value{
logical; TRUE or FALSE
}
\description{
The function tests wether a given data.frame has all important columns and the corresponding data types.
}
\examples{
is.RootdetectionOutput(root_test)

}
