% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{is_root_norm}
\alias{is_root_norm}
\title{Test Input data.frame for normalized Rootdetection standard}
\usage{
is_root_norm(root_norm)
}
\arguments{
\item{root_output}{data.frame; output *.csv (list) from Rootdetection}
}
\value{
logical; TRUE or FALSE
}
\description{
The function tests wether a given object fulfils all criteria of a normalized rootdetection dataset. These standards are:
- The object has to be a data.frame
- The columns Label, LengthMM and LengthPx, Factor1 and Factor2 needs to be present
- LengthMM needs to store numeric values
- No length standard must be present
- hyphens (-) in Labels are not allowed
}
\examples{
is_root_output(root_output)

}
