% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{norm_cust_standard}
\alias{norm_cust_standard}
\title{Calculate LengthMM from LengthPx for Rootdetection standard}
\usage{
norm_cust_standard(root_output, sort = TRUE, label_delim = ";",
  label_standard = "10mm", standard_length_mm = "10")
}
\arguments{
\item{root_output}{data.frame; *.csv output from Rootdetection containing 10mm values}

\item{sort}{logical; if TRUE data.frame is sorted and Label splitted in Factor1 and Factor2}

\item{label_delim}{character; defines how Factor1 and Factor2 are seperated in Label}

\item{label_standard}{string; defines  how the standard is labled in the data.frame}

\item{standard_length_mm}{numeric; defines the length (in mm) of the standard}
}
\value{
data.frame; containing normalized length values
}
\description{
This function takes customized standard values to calculate LengthMM from LengthPx. The label_standard must match the label in the data.frame. In addition the mm value of the standard must be provided.
}
\examples{
# use data_set with standard 10mm label
# add label_standard and standard_length_mm

norm_cust_standard(root_output, label_delim = ';', label_standard = '10mm', standard_length_mm = '10')

}
