% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{normality_test}
\alias{normality_test}
\title{Normality Test for Rootdetection standard}
\usage{
normality_test(root_norm)
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}
}
\value{
data.frame; containing p-values for each Factor1 Facto2 combinations.
}
\description{
The function performes a normality test for each Factor1 Factor 2 combination in Rootdetecion standard data.frame. Until now only Shapiro-Wilk test is implemented.
}
\examples{
root_test_norm <- norm_10mm_standard(root_test)
normality_test(root_test_norm)

}
