% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_hist}
\alias{plot_hist}
\title{Plotting Histograms}
\usage{
plot_hist(root_norm, draw_out = F, file = "data_distribution.pdf")
}
\arguments{
\item{root_norm}{data.frame; normalized Rootdetection data set}

\item{draw_out}{logical; TRUE = prints plot in pdf file, FALSE = no pdf output}

\item{file}{string; filename of the pdf output}
}
\value{
list; containg the p-values of shapiro-wilk test for each element
}
\description{
The function produces histogram plots for each grouping varibale 1 and 2 combinations (Factor1 and Factor2).
In addition a Shapiro-Wilk test of normaility is performed and p-value is plotted in the subtititle of the plots.
Also there is colour coding - green = normaly distributed according to Shapiro-Wilk test; red = not normaly distributed according to Shapiro-Wilk test
}
\examples{
# produce list of plots:

root_norm <- norm_10mm_standard(root_output)
plot_hist(root_norm, draw_out = F)
# view the first plot in R:
plot_hist[[1]]

# produce list of plots and print as pdf
root_norm <- norm_10mm_standard(root_output)
plot_hist(root_norm, draw_out = T, file = 'test_plot_histograms.pdf')
# view the first plot in R:
plot_hist[[1]]
}
