% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primroses}
\alias{primroses}
\title{Picture Of Colorful Primroses}
\format{A matrix containing color information for each pixel.}
\usage{
primroses
}
\description{
Picture of primroses loaded with 'jpeg' package. Picture made by 3268zauber~commonswiki and obtained from Wikimedia Commons.
}
\examples{
# show picture in R using grid package

grid.raster(primroses)
}
\keyword{datasets}
