% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{rel_data}
\alias{rel_data}
\title{Create Relative data from Rootdetection standard}
\usage{
rel_data(root_norm, control = "20")
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}
}
\value{
data.frame; containing data for each Factor2 relative to Factor2 control
}
\description{
The function takes the control values of Factor2 and calculates the relative data for the other Factor2
}
\examples{
root_test_norm <- norm_10mm_standard(root_test)
rel_data(root_test_norm)

}
