% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{rename_element}
\alias{rename_element}
\title{Change Names of Elements In Specific Column}
\usage{
rename_element(df, colname, old_name, new_name)
}
\arguments{
\item{df}{data.frame; input data.frame}

\item{colname}{character; name of the column storing the elements to be renamed}

\item{old_name}{vector; elements that should be renamed in column specified by 'colname'}

\item{new_name}{vector; new names of the elements specified by "old_name"}
}
\value{
data.frame; with renamed elements in "colname"
}
\description{
The function will rename elements in a specific column by vectors "old_name" and "new_name".
Simply give the function a vector with old names that should be renamed by a vector with new names.
_Be carefull that elements in new_name have the rigth position_ (e.g. the second element in "old_name" will be replaced by the second element in "new_name").
}
\examples{
rename_element(root_output, colname = 'Label', oldname = c('Col_0;28', 'Col_0;20'), new_name = c('Col0_28', 'Col0_20'))
}
