% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{rm_outlier}
\alias{rm_outlier}
\title{Remove Outlier}
\usage{
rm_outlier(x, fill_na = F)
}
\arguments{
\item{x}{numeric vector}

\item{fill.na}{logical; If TRUE all outliers present in x will be replaced by NA. If FALSE all outliers will be deleted.}
}
\value{
numeric vector; without outliers or outliers replaced by NA
}
\description{
This function removes outlier from a vector or replaced outliers by NA.
According to box plots the function defines outliers as 1.5 x IQR. Caution: NAs already present in the input vector will be removed first.
}
\examples{
# get some example vector
root_norm <- norm_10mm_standard(root_output)
x <- root_norm[root_norm$Label == "weitar1_1;28", ]

# delete outliers
rm_outlier(x, fill_na = F)

# replace outliers with NA
rm_outlier(x, fill_na = T)
}
