% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{rm_outlier_df}
\alias{rm_outlier_df}
\title{Remove Outlier from Rootdetection standard}
\usage{
rm_outlier_df(root_norm, fill_na = F)
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{fill_na}{logical; If TRUE all outliers present in x will be replaced by NA. If False all outliers will be deleted.}
}
\value{
data.frame; containing data without outliers or outliers replaced by NA
}
\description{
The function removes or replaces outliers for every Factor1 Factor2 (Label) combination. Outliers are defined as 1.5 x IQR.
}
\examples{
root_test_norm <- norm_10mm_standard(root_output)

# transform outliers to NA
rm_outlier_df(root_test_norm, fill_na = T)
# remove outliers
rm_outlier_df(root_test_norm, fill_na = F)

}
