% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootdetectR.R
\docType{package}
\name{rootdetectR}
\alias{rootdetectR}
\title{rootdetectR: Statistical analysis of rootdetection output with R}
\description{
rootdectR can be used to analyse the output from rootdetection software. This package contains functions to visualize and statistically test data coming from rootdetection
software. It can be used to produce pretty boxplots including statistics from ANOVA analysis.
}
\section{Included Data Sets}{

root_output, root_output_multfac2, primroses, colors_wong
}

\section{Data Processing Functions}{

is_root_output, is_root_norm, norm_10mm_standard, norm_cust_standard, summary_stat, normality_test, rel_data, rm_outlier_df
}

\section{Statistics (ANOVA) Functions}{

onefacaov_fac1, onefacaov_fac2, twofacaov, interaction_twofacaov
}

\section{Plotting Functions}{

plot_hist, plot_abs, plot_rel
}

\section{helpful Functions}{

se, rm_outlier, subset_fac2, tukey_to_matrix, get_sig_letters, sort_label, extract_colors, show_colors, detach_all, rename_element
}

