% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{se}
\alias{se}
\title{Calculate Standard Error}
\usage{
se(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
numeric value; standard error of vector x
}
\description{
This function calculates the standard error for a numeric-vector x. NA values will be removed.
}
\examples{
x <- 1:20
se(x)

}
