% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{show_colors}
\alias{show_colors}
\title{Show Color Palette}
\usage{
show_colors(x)
}
\arguments{
\item{x}{vector; must contain hex encoded or standard R colors}
}
\value{
plot; showing the colors in an example plot
}
\description{
Plotting a vector containing hex encoded or standard R colors. The function will return a plot showing all the colors that are stored in a vector.
}
\examples{
# show the color palette for colorblind individuals
show_colors(colors_wong)

}
