% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{subset_fac2}
\alias{subset_fac2}
\title{Create subsets by Factor2}
\usage{
subset_fac2(root_norm, control = "20")
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{control}{string; name of the Factor2 control condition}
}
\value{
list of data.frames; each subset will be in an seperated data.frame stored in a list of data.frames
}
\description{
This function splits Normalized Rootdetection Output in subsets for each Factor2 control and treatment pair.
To be functional Factor2 must contain a control and at least two other conditions (treatments)
}
