% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{summary_stat}
\alias{summary_stat}
\title{Calculate Summary Statistics for Rootdetection standard}
\usage{
summary_stat(root_norm)
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}
}
\value{
data.frame; containing summary statistics for each Factor1 - Factor2 combination
}
\description{
This function calculates some summary statistics for a LengthMM normalized Rootdetection output. Calculated values are: sample size (n), median, mean, standard deviation (sd), standard error (se)
}
\examples{
root_test_norm <- norm_10mm_standard(root_test)
summary_stat(root_test_norm)

}
