% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{tukey_to_matrix}
\alias{tukey_to_matrix}
\title{Create rootdetectR Compatible Matrix from TukeyHSD output}
\usage{
tukey_to_matrix(tukeyHSD_output)
}
\arguments{
\item{tukeyHSD_output}{matrix or data.frame; Output from the TukeyHSD() function.}
}
\value{
matrix; containing the p-values from TukeyHSD() in an appropriate format for other rootdetectR functions.
}
\description{
The function will create a rootdetectR compatible Matrix from TukeyHSD output. The returened matrix can be used to retrieve plotting letters or can be used directly inside the plotting functions (plot_rel, plot_abs).
}
\examples{
# conduct ANOVA and TukeyHSD
root_norm <- norm_10mm_standard(root_output)
aov_all_vs_all <- aov(LengthMM ~ Factor1 * Factor2, data = root_norm)
tuk <- TukeyHSD(aov_all_vs_all, ordered = FALSE)$`Factor1:Factor2`

# use Tukey Output to retrieve matrix
tukey_to_matrix(tuk)

}
