% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{twofacaov}
\alias{twofacaov}
\title{Two Way ANOVA}
\usage{
twofacaov(root_norm, grouping_var1 = "Factor1",
  grouping_var2 = "Factor2", dependend_var = "LengthMM",
  label_delim = ";", summary_plots = F, draw_out = F,
  file = "2fac_ANOVA_all_vs_all.pdf", axis_label_size = 0.7,
  p_value_size = 0.8)
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{grouping_var1}{string; column name of the first grouping variable (Factor1)}

\item{grouping_var2}{string; column name of the second grouping variable (Factor2)}

\item{dependend_var}{string; column name of the dependend variable}

\item{label_delim}{character; defin how Factor1 and Factor2 are seperated in Label}

\item{summary_plots}{logical; If TRUE summary plots were printed (plot(aov))}

\item{draw_out}{logical; If TRUE Matrix containg p-values is plotted in pdf file}

\item{file}{string; file name of the pdf output (is needed if draw_out = T)}

\item{axis_label_size}{numeric; font size of axis labels}

\item{p_value_size}{numeric; font size of the p-values printed in pdf file}
}
\value{
list; data.frames containg p-values for one way ANOVA over Factor2
}
\description{
The function performes a two way ANOVA for Rootdetection standard
}
\examples{
# get data.frame containg p-values for one way ANOVA over Factor 2

root_test_norm <- norm_10mm_standard(root_output)
twofacaov(root_test_norm, label_delim = ';', draw_out = F)

# get data.frame and plot as pdf output

root_test_norm <- norm_10mm_standard(root_output)
twofacaov(root_test_norm, label_delim = ';', draw_out = T, file = '2fac_ANOVA_all_vs_all.pdf')
# change size of printed p-values
twofacaov(root_test_norm, label_delim = ';', draw_out = T, file = '2fac_ANOVA_all_vs_all.pdf', p_value_size = 1)
}
