% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{is_root_output}
\alias{is_root_output}
\title{Test Input Data.frame For Rootdetection Output}
\usage{
is_root_output(root_output, length_standard = "10mm")
}
\arguments{
\item{root_output}{data.frame; output *.csv (list) from Rootdetection}
}
\value{
logical; TRUE or FALSE
}
\description{
The function tests whether a given object fulfills all criteria of a standard Rootdetection output. These standards are:

- The object has to be a data.frame

- The columns Label and LengthPx needs to be present

- LengthPx needs to store numeric values

- A length standard needs to be present (default = 10mm) - can be changed by argument length_standard

- hyphens (-) in Labels are not allowed
}
\examples{
is_root_output(root_output)
}
