% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{norm_10mm_standard}
\alias{norm_10mm_standard}
\title{Calculate LengthMM from LengthPx for Rootdetection standard}
\usage{
norm_10mm_standard(root_output, sort = TRUE, label_delim = ";")
}
\arguments{
\item{root_output}{data.frame; *.csv output from Rootdetection containing 10mm values}

\item{sort}{logical; if TRUE data.frame is sorted and Label splitted in Factor1 and Factor2}

\item{label_delim}{character; defines how Factor1 and Factor2 are seperated in Label}
}
\value{
data.frame; containing normalized length values
}
\description{
This function takes 10mm standard values to calculate LengthMM from LengthPx
}
\examples{
norm_10mm_standard(root_output)
norm_10mm_standard(root_output, sort = FALSE)

}
