% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{normality_test}
\alias{normality_test}
\title{Normality Test For Normalized Rootdetection Data Set}
\usage{
normality_test(root_norm, col_grouping = "Label", col_value = "LengthMM")
}
\arguments{
\item{root_norm}{data.frame; normalized Rootdetection data set}

\item{col_grouping}{string; name of the column that should be used as grouping variable (Label)}

\item{col_value}{string; name of the column containing values (dependent variable) (LengthMM)}
}
\value{
data.frame; containing p-values for each grouping variable
}
\description{
The function performs a Shapiro-Wilk test of normality of dependent variable (LengthMM) for each grouping variable (Label).
A data.frame containing p-values obtained from the Shapiro-Wilk test for each grouping variable (Label) is returned.
}
\examples{
# normalize Rootdetection Output
root_norm <- norm_10mm_standard(root_output)

normality_test(root_norm)
}
