% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{onefacaov_fac1}
\alias{onefacaov_fac1}
\title{1 Factorial ANOVA Over Factor1}
\usage{
onefacaov_fac1(root_norm, draw_out = T,
  file_base = "1fac_ANOVA_factor1")
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{draw_out}{logical; If TRUE Matrix containg p-values is plotted in pdf file}

\item{file_base}{string; file base name of the pdf output (is needed if draw_out = T)}
}
\value{
list; data.frames containg p-values for one way ANOVA over Factor 1
}
\description{
The function performes a one way ANOVA over Factor1. The function iterates over Factor 2.
}
\examples{
# get data.frame containg p-values for one way ANOVA over Factor 1

root_test_norm <- norm_10mm_standard(root_test)
onefacaov_fac1(root_test_norm, draw_out = F)

# get data.frame and plot as pdf output
root_test_norm <- norm_10mm_standard(root_test)
onefacaov_fac1(root_test_norm, draw_out = T, file_base = '1fac_ANOVA_factor1')
# create two pdf files: 1fac_ANOVA_factor1_20.pdf and 1fac_ANOVA_factor1_28.pdf
}
