% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{onefacaov_fac2}
\alias{onefacaov_fac2}
\title{One-way ANOVA Over Grouping Variable 2 Treatments}
\usage{
onefacaov_fac2(
  root_norm,
  col_grouping1 = "Factor1",
  col_grouping2 = "Factor2",
  col_value = "LengthMM",
  control = "20",
  draw_out = F,
  file_base = "1fac_ANOVA_factor2",
  axis_label_size = 0.7,
  p_value_size = 0.8
)
}
\arguments{
\item{root_norm}{data.frame; normalized Rootdetection data set}

\item{col_grouping1}{string; name of the column that should be used as grouping variable 1 (Factor1)}

\item{col_grouping2}{string; name of the column that should be used as grouping variable 2 (Factor2)}

\item{col_value}{string; name of the column containing values (dependent variable) (LengthMM)}

\item{control}{string; name of the grouping variable 2 (Factor2) control condition}

\item{draw_out}{logical; If TRUE a matrix containing p-values is plotted in a pdf file}

\item{file_base}{string; file base name of the pdf output (is needed if draw_out = TRUE)}

\item{axis_label_size}{numeric; font size of axis labels in pdf file (if draw_out = TRUE)}

\item{p_value_size}{numeric; font size of p-values in pdf file (if draw_out = TRUE)}
}
\value{
list; data.frames containing p-values for one-way ANOVA and Tukey post-hoc
}
\description{
The function performs a one-way ANOVA and Tukey post-hoc test over grouping variable 2 (Factor2).
The function takes each control + treatment combination of grouping variable 2 and perfoms a one-way ANOVA
iterating over grouping Variable 1 (Factor1).
The p-values are adjusted using Benjamini-Hochberg correction.
}
\examples{
# get data.frame containg p-values for one-way ANOVA and Tukey post-hoc over grouping variable 2
root_norm <- norm_10mm_standard(root_output)
onefacaov_fac2(root_norm, control = "20", draw_out = FALSE)

# plot p-value matrix as pdf output
onefacaov_fac2(root_norm, control = "20", draw_out = TRUE, file_base = "1fac_ANOVA_factor2")
# function creates a pdf file 1fac_ANOVA_factor2_28.pdf
}
