% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_abs}
\alias{plot_abs}
\title{Plotting absolute data of Rootdetection standard}
\usage{
plot_abs(root_norm, plot_significance = F, twofacov_output,
  label_delim = ";", type = "jitter", plot_colours,
  letter_height = 2)
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{plot_significance}{logical; if TRUE significances will be drawn as letters}

\item{label_delim}{character; defin how Factor1 and Factor2 are seperated in Label}

\item{type}{string; "box" = will produce Boxplot, 'jitter' = will produce combination of box and jitter plot}

\item{plot_colours}{vector; provide colours for the boxplot - colour vector must have the same length than Factor2}

\item{letter_height}{numeric; defines the position of the significance letters}

\item{twofacaov_output}{data.frame; Output of twofacaov (needed if plot_significance = T)}
}
\value{
plot; box or box-jitter-plot of the absolute data
}
\description{
Absolute data are plotted as boxplot or box-jitter-plot combination. Signifcances can be illustrated in the plot.
}
\examples{
# Plot without significance

root_test_norm <- norm_10mm_standard(root_test)
# boxplot
plot_abs(root_test_norm, plot_significance = F, type = 'box')
# jitterplot
plot_abs(root_test_norm, plot_significance = F, type = 'jitter')

# Plot with siginficance letters
root_test_norm <- norm_10mm_standard(root_test)
root_stat <- twofacaov(root_test_norm, draw_out = F)
# boxplot with statistics
plot_abs(root_test_norm, plot_significance = T, twofacaov_output = root_stat, type = 'box')

# provide own colours (ggplot2-colours or hex-code) - must have the same length than level(root_test_norm$Factor2)
root_test_norm <- norm_10mm_standard(root_test)
plot_abs(root_test_norm, plot_significance = F, type = 'jitter', plot_colours = c('dodgerblue3', 'firebrick2'))

}
