% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_hist}
\alias{plot_hist}
\title{Plotting histograms and performing normailty tests}
\usage{
plot_hist(root_norm, draw_out = T, file = "data_distribution.pdf")
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{draw_out}{logical; TRUE = prints plot in pdf file, False = no pdf output}

\item{file}{string; filename of the pdf output}
}
\value{
list; containg the p-values of shapiro-wilk test for each element
}
\description{
The function performes a shapiro-wilk normality test for each Factor1 Factor 2 combination in Rootdetecion standard data.frame and plots a histogram for each containing the p-value from shaprio-wilk test and colour coding - green = normaly distributed; red = not normaly distributed.
}
\examples{
#Only produce list of plots:

root_test_norm <- norm_10mm_standard(root_output)
plot_hist(root_test_norm, draw_out = F)
#To view the first plot in R:
plot_hist[[1]]

#Produce plots and print as pdf

root_test_norm <- norm_10mm_standard(root_output)
plot_hist(root_test_norm, draw_out = T, file = 'test_plot_histograms.pdf')
#To view the first plot in R:
plot_hist[[1]]

}
