% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rel}
\alias{plot_rel}
\title{Plotting relative data of Rootdetection standard}
\usage{
plot_rel(root_norm, plot_significance = T, pairwise_2facaov_output,
  control = "20", type = "jitter", letter_height = 10)
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{plot_significance}{logical; if TRUE significances will be drawn as letters}

\item{pairwise_2facaov_output}{data.frame; Output of pairwise_2facaov (needed if plot_significance = T)}

\item{control}{string; name of the Factor2 control condition}

\item{type}{string; "box" = will produce Boxplot, 'jitter' = will produce combination of box and jitter plot}

\item{letter_height}{numeric; defines the position of the significance letters}
}
\value{
plot or list of plots; box or box-jitter-plot of relative data, if significance = T and multiple Factor2 treatments a list of boxplots will be generated
}
\description{
Relative data are plotted as boxplot or box-jitter-plot combination. If Significances should be illustrated multiple plots are generated. Each Factor2 control Factor2 treatment combination will produce a plot.
}
\examples{
# Plot without significance

root_test_norm <- norm_10mm_standard(root_output)
# boxplot
plot_rel(root_test_norm, plot_significance = F, control = '20', type = 'box')
# jitterplot
plot_rel(root_test_norm, plot_significance = F, control = '20', type = 'jitter')

# Plot with siginficance letters

root_test_norm <- norm_10mm_standard(root_output)
root_stat <- pairwise_2facaov(root_test_norm, draw_out = F, label_delim = ';')
# boxplot with statistics
plot_rel(root_test_norm, plot_significance = T, pairwise_2facaov_output = root_stat, control = '20', type = 'box')

}
