% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rel}
\alias{plot_rel}
\title{Plotting Relative Data Of Normalized Rootdetection Data Set}
\usage{
plot_rel(
  root_norm,
  label_delim = ";",
  control = 20,
  type = "jitter",
  size_jitter_dot = 2,
  plot_colours,
  plot_title,
  size_plot_title = 14,
  y_label = "hypocotyl growth [\%]",
  x_label = "",
  legend_label = "Label",
  size_legend_title = 12,
  size_legend_text = 10,
  width_lines = 0.5,
  width_axis = 0.5,
  size_x_axes = 9,
  size_y_axes = 9,
  plot_significance = F,
  height_letter = 10,
  size_letter = 5,
  angle_letter = 0
)
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{label_delim}{character; define how Factor1 and Factor2 are seperated in Label}

\item{control}{string; name of the Factor2 control condition}

\item{type}{string; "box" = will produce Boxplot, 'jitter' = will produce combination of box and jitter plot, 'violin' = will produce violin plot}

\item{size_jitter_dot}{number; defines the size of the dots in jitter plots}

\item{plot_colours}{vector; provide colours for the boxplot - colour vector must have the same length than Factor1}

\item{plot_title}{string; defines the plot title}

\item{size_plot_title}{numeric; defines size of the plot title}

\item{y_label}{string; axes description of y-axes}

\item{x_label}{string; axes description of x-axes}

\item{legend_label}{string; title of legend}

\item{size_legend_title}{numeric; defines size of legend title}

\item{size_legend_text}{numeric; defines size of legend text}

\item{width_lines}{numeric; defines the line width of boxes or violins}

\item{width_axis}{numeric; defines the line width of the axis}

\item{size_x_axes}{numeric; defines size of x-axes labels}

\item{size_y_axes}{numeric; defines size of y-axes labels}

\item{plot_significance}{logical; if TRUE significances will be drawn as letters}

\item{height_letter}{numeric; defines the position of the significance letters}

\item{size_letter}{numeric; defines size of the significance letters}

\item{angle_letter}{numeric, defines angle of significance letters}
}
\value{
plot or list of plots; box or box-jitter-plot of relative data, if significance = T and multiple Factor2 treatments a list of boxplots will be generated
}
\description{
Relative data is plotted as boxplot or box-jitter-plot combination. If Significances should be illustrated multiple plots are generated. Each Factor2 control Factor2 treatment combination will produce a plot.
}
\examples{
# Plot without significance

root_test_norm <- norm_10mm_standard(root_output)
# boxplot
plot_rel(root_test_norm, plot_significance = F, control = "20", type = "box")
# jitterplot
plot_rel(root_test_norm, plot_significance = F, control = "20", type = "jitter")

# Plot with siginficance letters

root_test_norm <- norm_10mm_standard(root_output)
root_stat <- interaction_twofacaov(root_test_norm, draw_out = F, label_delim = ";")
# boxplot with statistics
plot_rel(root_test_norm, plot_significance = T, interaction_twofacaov_output = root_stat, control = "20", type = "box")

# all customizable plotting parameters have a default value
plot_rel(root_test_norm,
  label_delim = ";",
  control = 20,
  type = "jitter",
  plot_colours = c("blue", "red", "orange", "green"),
  y_label = "\% growth",
  x_label = "",
  legend_label = "Label",
  size_legend_title = 12,
  size_legend_text = 10,
  size_x_axes = 9,
  size_y_axes = 9,
  plot_significance = T,
  interaction_twofacaov_output = root_stat,
  height_letter = 10,
  size_letter = 5,
  angle_letter = 0
)

}
