% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{rm_outlier}
\alias{rm_outlier}
\title{Remove Outlier}
\usage{
rm_outlier(x, fill_na = F)
}
\arguments{
\item{x}{numeric vector}

\item{fill.na}{logical; If TRUE all outliers present in x will be replaced by NA. If False all outliers will be deleted.}
}
\value{
numeric vector; without outliers or outliers replaced by NA
}
\description{
This function removes outlier from a vector and returns either a vector without the outliers or replaces the outliers with NA.
Outliers are defined as 1.5 x IQR. Caution: NAs already present in the input vector will be removed.
}
\examples{
# get some example vector
root_test_norm <- norm_10mm_standard(root_output)
test_vector <- root_test_norm[root_test_norm$Label == 'weitar1_1;28',]

# just delete outliers
rm_outlier(test_vector, fill_na = F)

# replace outliers with NA
rm_outlier(test_vector, fill_na = T)

}
