% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{root_output}
\alias{root_output}
\title{Rootdetection Output of an example study}
\format{A data frame with 125 rows of  6 variables:
\describe{
  \item{Nr}{Number given by Rootdetection}
  \item{Filename}{Filename of the photo}
  \item{RootNr}{Number given by Rootdetection}
  \item{Label}{Label describing the line}
  \item{LengthPx}{Length of measured root or Hypocotyl in Pixel}
  \item{LengthMM}{Length of measured root or Hypocotyl in mm}
}}
\usage{
root_output
}
\description{
A dataset containing measured A. thaliana wild-type (Col-0) and mutant (tir1_1afb2_3, weitar1_1, yucOx) seedling hypocotyls and a 10mm length standard in colum LengthPx.
}
\keyword{datasets}
