% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{root_output}
\alias{root_output}
\title{Rootdetection Output Of An Example Study}
\format{A data frame with 125 rows of  6 variables:
\describe{
  \item{Nr}{Number given by Rootdetection}
  \item{Filename}{Filename of the photo}
  \item{RootNr}{Number given by Rootdetection}
  \item{Label}{Label describing the line}
  \item{LengthPx}{Length of measured root or Hypocotyl in Pixel}
  \item{LengthMM}{Length of measured root or Hypocotyl in mm}
}}
\usage{
root_output
}
\description{
A data set containing measured _Arabidopsis thaliana_ wild-type (Col-0) and mutant (tir1_1afb2_3, weitar1_1, yucOx) seedling roots measured at 20 °C and 28°C.
Plants were grown on ATS plates for 4 days at 20°C and then stayed at 20°C or were shifted to 28°C.
The rows labeled with '10mm' in column Label contain the length standard in column LengthPx.
}
\keyword{datasets}
