% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{root_output_multfac2}
\alias{root_output_multfac2}
\title{Rootdetection Output of an example study with multiple Factor2}
\format{A data frame with 213 rows of  6 variables:
\describe{
  \item{Nr}{Number given by Rootdetection}
  \item{Filename}{Filename of the photo}
  \item{RootNr}{Number given by Rootdetection}
  \item{Label}{Label describing the line}
  \item{LengthPx}{Length of measured root or Hypocotyl in Pixel}
  \item{LengthMM}{Length of measured root or Hypocotyl in mm}
}}
\usage{
root_output_multfac2
}
\description{
A dataset containing measured _Arabidopsis thaliana_ wild-type (Col-0 and Ws2) and mutant (D31 and D64) seedling roots measured at 20 °C and 28°C. One part of the plants was grown at 20°C or 28°C for 8 days. Another batch was transferred after 4 days from 20°C to 28°C for another 4 days.
The columns labeled with '20mm' contains the length standard in colum LengthPx. In this dataset a different length standard is used (2cm are measured and are labled as 20mm).
}
\keyword{datasets}
