% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{sort_label}
\alias{sort_label}
\title{Sort Labels of Normalized Rootdetection Table}
\usage{
sort_label(root_norm, col_label = "Label", control_fac1, control_fac2)
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{col_label}{string; name of the column carring the labels (grouping variable/s)}

\item{control_fac1}{string; name of the control condition of grouping variable1 (Factor1)}

\item{control_fac2}{string; name of the control condition of grouping variable2 (Factor2)}
}
\value{
data.frame; with sorted col_label
}
\description{
The function will sort the labels according to Controls. By defining control_fac1 and control_fac2 you can set those conditions to the first place in order.
}
\examples{
# Obtain normalized Rootdetection data.frame
root_norm <- norm_10mm_standard(root_output)

# check the labels
levels(root_norm$Label)

# sort setting yucOx and 28 as controls
sort_label(root_norm, control_fac1 = 'yucOx', control_fac2 = '28')


}
