% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{subset_fac2}
\alias{subset_fac2}
\title{Create Subsets Of Normalized Rootdetection Standard Data Set By Factor2}
\usage{
subset_fac2(root_norm, control = "20")
}
\arguments{
\item{root_norm}{data.frame; normalized Rootdetection data set}

\item{control}{string; name of the grouping variable 2 (Factor2) control condition}
}
\value{
list of data.frames; each subset will be in an separated data.frame stored in a list
}
\description{
This function creates subsets for each grouping variable 2 (Factor2) control and treatment pair of a normalized Rootdetection data set.
(At least two treatment conditions are required)
}
\examples{
# create normalized data.set with multiple Factor2
root_norm_multfac2 <- norm_cust_standard(root_output_multfac2, label_standard = '20mm', standard_length_mm = '20')

# create subsets for each control_treatment pair
subset_fac2(root_norm_multfac2, control = '20')
}
