% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{twofacaov}
\alias{twofacaov}
\title{Two Way ANOVA}
\usage{
twofacaov(root_norm, label_delim = ";", draw_out = T,
  file = "2fac_ANOVA_all_vs_all.pdf")
}
\arguments{
\item{root_norm}{data.frame; LengthMM normalized output from Rootdetection containing NO 10mm values}

\item{label_delim}{character; defin how Factor1 and Factor2 are seperated in Label}

\item{draw_out}{logical; If TRUE Matrix containg p-values is plotted in pdf file}

\item{file}{string; file name of the pdf output (is needed if draw_out = T)}
}
\value{
list; data.frames containg p-values for one way ANOVA over Factor2
}
\description{
The function performes a two way ANOVA for Rootdetection standard
}
\examples{
# get data.frame containg p-values for one way ANOVA over Factor 2

root_test_norm <- norm_10mm_standard(root_test)
twofacaov(root_test_norm, label_delim = ';', draw_out = F)

# get data.frame and plot as pdf output
root_test_norm <- norm_10mm_standard(root_test)
twofacaov(root_test_norm, label_delim = ';', draw_out = T, file = '2fac_ANOVA_all_vs_all.pdf')
}
